feat(devtools): shared redaction module + dev-mode window bridge
Extract `redactSensitiveLeaves`, `REDACTED`, and the new
`DEVTOOLS_WINDOW_KEY` / `AttaformDevtoolsBridge` types into
`src/runtime/core/devtools-shared.ts`. The Vue DevTools wire-up in
`devtools.ts` now imports redaction from the shared module instead of
inlining it — one policy, one file, two consumers (Vue extension +
new Nuxt overlay panel).

The Nuxt runtime plugin attaches the registry to
`window.__attaform_devtools__` in dev mode so the overlay panel
(running in an iframe at /_attaform_devtools) can reach the host
app's reactive registry via `window.parent`. Vue's reactivity flows
across the same-origin iframe boundary cleanly; the bridge is just a
referent, not a serialised snapshot. Tree-shaken from production
builds via the `import.meta.dev` gate.

Re-export the shared primitives from the public `attaform` entry so
the panel's `.vue` files (shipped under `dist/runtime/`) can route
runtime imports through `'attaform'` rather than brittle relative
paths into the rollup-bundled chunk layout.
