+++ src/runtime/core/devtools.ts
@@ -1,9 +1,10 @@
 import type { App } from 'vue'
 import type { FormStore } from './create-form-store'
+import { redactSensitiveLeaves } from './devtools-shared'
 import type { AttaformRegistry } from './registry'
 import type { GenericForm } from '../types/types-core'
 import type { FormKey } from '../types/types-api'
-import { canonicalizePath, type Segment } from './paths'
+import { canonicalizePath } from './paths'
 
 /**
  * Vue DevTools plugin wiring for attaform. Lazy-imported by
@@ -26,72 +27,6 @@ import { canonicalizePath, type Segment } from './paths'
 
 const INSPECTOR_ID = 'attaform'
 const TIMELINE_LAYER_ID = 'attaform:events'
-
-const REDACTED = '[redacted]'
-
-/**
- * Walk `value` and replace any leaf whose enclosing path matches the
- * sensitive-name heuristic with the string `'[redacted]'`. Returns a
- * new tree (no mutation of the input). Object keys + array indices
- * are preserved; only the leaf payloads change.
- *
- * Applied to BOTH the DevTools timeline events and the inspector
- * `Form value` panel — leaks via either surface are treatable as
- * "any developer with the panel open during user testing can read
- * a customer's password," which is exactly the failure mode the
- * sensitive-name guard exists to prevent on the storage side.
- *
- * Leaves whose path doesn't match a pattern pass through untouched.
- * `acknowledgeSensitive: true` on persistence does NOT bypass this —
- * if the consumer opted into persisting the value, they still
- * shouldn't see it in DevTools timelines that grow unbounded.
- *
- * Implementation note: tracks an `inSensitiveSubtree` flag through
- * the recursion instead of allocating a fresh path array per node
- * + calling `isSensitivePath` per leaf. Once any ancestor segment
- * matches the heuristic, the flag stays set for every descendant —
- * the leaf simply returns `REDACTED` without re-scanning the path.
- * For a 100-leaf form: ~100 path allocations + ~100 full-path regex
- * sweeps → 0 path allocations + ~100 single-segment regex sweeps,
- * with whole-subtree short-circuit when sensitive ancestors are
- * found early.
- */
-function redactSensitiveLeaves(
-  value: unknown,
-  matchSensitive: (segment: Segment) => boolean
-): unknown {
-  return redactImpl(value, false, matchSensitive)
-}
-
-function redactImpl(
-  value: unknown,
-  inSensitiveSubtree: boolean,
-  matchSensitive: (segment: Segment) => boolean
-): unknown {
-  if (value === null || value === undefined) return value
-  if (typeof value !== 'object') {
-    return inSensitiveSubtree ? REDACTED : value
-  }
-  if (Array.isArray(value)) {
-    // Numeric segments never match the sensitive-name heuristic
-    // (segmentMatchesSensitive rejects non-string segments), so the
-    // flag passes through unchanged when descending into arrays.
-    return value.map((item) => redactImpl(item, inSensitiveSubtree, matchSensitive))
-  }
-  // Non-plain object (Map / Set / Date / class instance) — redact
-  // wholesale if we're already in a sensitive subtree; otherwise pass
-  // through. DevTools rendering of these is already heuristic, so we
-  // don't try to descend into them.
-  if (Object.getPrototypeOf(value) !== Object.prototype && Object.getPrototypeOf(value) !== null) {
-    return inSensitiveSubtree ? REDACTED : value
-  }
-  const out: Record<string, unknown> = {}
-  for (const key of Object.keys(value as Record<string, unknown>)) {
-    const childSensitive = inSensitiveSubtree || matchSensitive(key)
-    out[key] = redactImpl((value as Record<string, unknown>)[key], childSensitive, matchSensitive)
-  }
-  return out
-}
 
 type UnsafeDevtoolsApi = {
   addInspector(opts: { id: string; label: string; icon?: string; app: App }): void
